Removed tests that were not supposed to be there, combined the behavior test classes and removed service provider as dependency
