+++ src/Workleap.DomainEventPropagation.Tests/Publishing/EventPropagationClientTests.cs
@@ -9,7 +9,6 @@ namespace Workleap.DomainEventPropagation.Tests.Publishing;
 
 public class EventPropagationClientTests
 {
-    private readonly IServiceProvider _serviceProvider;
     private readonly EventPropagationPublisherOptions _eventPropagationPublisherOptions;
     private readonly EventPropagationClient _eventPropagationClient;
     private readonly IAzureClientFactory<EventGridPublisherClient> _eventGridPublisherClientFactory;
@@ -22,12 +21,11 @@ public class EventPropagationClientTests
         this._eventPropagationPublisherOptions = new EventPropagationPublisherOptions { TopicName = "Organization", TopicAccessKey = "AccessKey", TopicEndpoint = "http://localhost:11111" };
         this._eventGridPublisherClientFactory = A.Fake<IAzureClientFactory<EventGridPublisherClient>>(opts => opts.Strict());
         this._eventGridPublisherClient = A.Fake<EventGridPublisherClient>(opts => opts.Strict());
-        this._serviceProvider = A.Fake<IServiceProvider>();
 
         this._eventPropagationClient = new EventPropagationClient(
             this._eventGridPublisherClientFactory,
             Options.Create(this._eventPropagationPublisherOptions),
-            this._serviceProvider);
+            Array.Empty<IPublishingDomainEventBehavior>());
     }
 
     [Fact]
@@ -70,7 +68,6 @@ public class EventPropagationClientTests
     public async Task GivenEventPropagationClient_WhenEventsAreSuccessfullySentWithEventGridPublisher_ThenNoErrors()
     {
         A.CallTo(() => this._eventGridPublisherClientFactory.CreateClient(EventPropagationPublisherOptions.ClientName)).Returns(this._eventGridPublisherClient);
-        A.CallTo(() => this._serviceProvider.GetService(A<Type>._)).Returns(Array.Empty<IPublishingDomainEventBehavior>());
 
         A.CallTo(() => this._eventGridPublisherClient.SendEventsAsync(
                 A<IEnumerable<EventGridEvent>>.That.Matches(events => events.Count() == 1),
@@ -84,7 +81,6 @@ public class EventPropagationClientTests
     public async Task GivenGenericPublishDomainEventsAsync_WhenEventsAreSuccessfullySentWithEventGridPublisher_ThenNoErrors()
     {
         A.CallTo(() => this._eventGridPublisherClientFactory.CreateClient(EventPropagationPublisherOptions.ClientName)).Returns(this._eventGridPublisherClient);
-        A.CallTo(() => this._serviceProvider.GetService(A<Type>._)).Returns(Array.Empty<IPublishingDomainEventBehavior>());
 
         var domainEvents = new List<PublishTestDomainEvent> { this._domainEvent, new() };
 
@@ -100,7 +96,6 @@ public class EventPropagationClientTests
     public async Task GivenGenericPublishDomainEventsAsync_WhenEventsAreSuccessfullySentWithEventGridPublisher_DataVersion_ShouldBe_1_0()
     {
         A.CallTo(() => this._eventGridPublisherClientFactory.CreateClient(EventPropagationPublisherOptions.ClientName)).Returns(this._eventGridPublisherClient);
-        A.CallTo(() => this._serviceProvider.GetService(A<Type>._)).Returns(Array.Empty<IPublishingDomainEventBehavior>());
 
         var domainEvents = new List<PublishTestDomainEvent> { this._domainEvent, new() };
 
@@ -118,10 +113,6 @@ public class EventPropagationClientTests
         // Given
         var publisherBehavior = A.Fake<IPublishingDomainEventBehavior>();
 
-        var services = new ServiceCollection();
-        services.AddSingleton(publisherBehavior);
-        var serviceProvider = services.BuildServiceProvider();
-
         var publisherClient = A.Fake<EventGridPublisherClient>();
         var clientFactory = A.Fake<IAzureClientFactory<EventGridPublisherClient>>();
         A.CallTo(() => clientFactory.CreateClient(EventPropagationPublisherOptions.ClientName)).Returns(publisherClient);
@@ -129,7 +120,7 @@ public class EventPropagationClientTests
         var propagationClient = new EventPropagationClient(
             clientFactory,
             Options.Create(this._eventPropagationPublisherOptions),
-            serviceProvider);
+            new[] { publisherBehavior });
 
         // When
         await propagationClient.PublishDomainEventAsync(this._domainEvent, CancellationToken.None);
