+++ src/Workleap.DomainEventPropagation.Tests/PublishingDomainEventTracingBehaviorTests.cs
@@ -1,18 +1,28 @@
 ﻿using Azure.Messaging.EventGrid;
 using GSoft.Extensions.Xunit;
 using Microsoft.Extensions.DependencyInjection;
-using Workleap.DomainEventPropagation.Tests.Publishing;
 
-namespace Workleap.DomainEventPropagation.Tests.Subscription;
+namespace Workleap.DomainEventPropagation.Tests;
 
-public sealed class SubscribtionDomainEventTracingBehaviorTests : BaseUnitTest<TracingFixture>
+public sealed class PublishingDomainEventTracingBehaviorTests : BaseUnitTest<TracingBehaviorFixture>
 {
     private readonly InMemoryActivityTracker _activities;
+    private readonly IEventPropagationClient _eventPropagationClient;
 
-    public SubscribtionDomainEventTracingBehaviorTests(TracingFixture fixture, ITestOutputHelper testOutputHelper)
+    public PublishingDomainEventTracingBehaviorTests(TracingBehaviorFixture fixture, ITestOutputHelper testOutputHelper)
         : base(fixture, testOutputHelper)
     {
+        this._eventPropagationClient = this.Services.GetRequiredService<IEventPropagationClient>();
         this._activities = this.Services.GetRequiredService<InMemoryActivityTracker>();
+    }
+
+    [Fact]
+    public async Task GivenActivityListener_WhenPublishDomainEvent_ThenHandleWithTracing()
+    {
+        var domainEvent = new SampleDomainEvent();
+        await this._eventPropagationClient.PublishDomainEventAsync(domainEvent, CancellationToken.None);
+
+        this._activities.AssertPublishSuccessful();
     }
 
     [Fact]
